Move logic to hooks
