+++ src/client/hooks/useRouteParams.ts
@@ -1,6 +1,12 @@
 import { useParams } from 'react-router-dom'
 
-import { AssessmentRouteParams, CountryRouteParams, CycleRouteParams, SectionRouteParams } from 'meta/routes'
+import {
+  AssessmentRouteParams,
+  CountryRouteParams,
+  CycleRouteParams,
+  OriginalDataPointRouteParams,
+  SectionRouteParams,
+} from 'meta/routes'
 
 export const useAssessmentRouteParams = () => useParams<AssessmentRouteParams>()
 
@@ -9,3 +15,5 @@ export const useCycleRouteParams = () => useParams<CycleRouteParams>()
 export const useCountryRouteParams = () => useParams<CountryRouteParams>()
 
 export const useSectionRouteParams = () => useParams<SectionRouteParams>()
+
+export const useOriginalDataPointRouteParams = () => useParams<OriginalDataPointRouteParams>()
